editing cabin

+++ advanced-react/apps/08-the-wild-oasis/src/services/apiCabins.js
@@ -20,29 +20,42 @@ export async function deleteCabin(id) {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     "",
   );
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
   try {
-    const { data, error } = await supabase
-      .from("cabins")
-      .insert([{ ...newCabin, image: imagePath }])
-      .select();
+    // 1. Create/edit cabin
+    let query = supabase.from("cabins");
+    // A) CREATE
+    if (!id) {
+      query = query.insert([{ ...newCabin, image: imagePath }]);
+    }
+    // B) EDIT
+    if (id) {
+      query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+    }
+
+    const { data, error } = await query.select().single();
 
     if (error) {
       throw new Error("Cabin could not be added");
     }
 
     // 2 upload image
+    if (hasImagePath) return data;
 
     const { error: storageError } = await supabase.storage
       .from("cabin-images")
       .upload(imageName, newCabin.image);
 
+    // 3. Delete the cabin IF there was an error uplaoding image
     if (storageError) {
       await supabase.from("cabins").delete().eq("id", data.at(0).id);
       console.error(storageError);
